+++ RSSBot/pylib/rssNews/international/france24/france24World.py
@@ -1,0 +1,306 @@
+
+#   Python Responsories
+
+#   Discord Responsories
+
+from discord.utils import get
+from discord.embeds import Embed
+from discord.colour import Color
+from discord.ext.commands import Cog, command
+from discord.permissions import PermissionOverwrite
+
+#   RSS Responsories
+import feedparser
+
+class France24World(Cog):
+    def __init__(self, bot) -> None:
+        self.channelName = 'rssfeed'
+        self.embed = Embed(color=Color.dark_blue())
+
+
+    @command(name='24world')
+    async def France24World(self, ctx):
+
+            #   Retrieve the guild information
+        srv = ctx.guild
+        member = get(srv.roles, name='@Members')
+        ch = get(srv.channels, name=f'{self.channelName}')
+        rss = 'https://www.france24.com/en/rss'
+
+        #   Creating the channel if it does not exists
+        if not ch:
+
+            #   Inserting the channel permissions
+            perms = {
+                        srv.default_role:PermissionOverwrite(read_messages=False),
+                        member:PermissionOverwrite(view=True, read_channel_history=True)
+}
+
+            await srv.create_text_channel(f'{self.channelName}', overwrites=perms)
+
+        #   Creating the RSS 
+        RSSNews = feedparser.parse(f'{rss}')
+        entries = RSSNews.entries
+
+        #   Prepareing the embed
+        self.embed.title = f'France 24 World'
+        self.embed.description = f''
+
+
+        #   Looping throught the entries
+        for nr, article in enumerate(entries):
+
+            summary = article.get('summary', 'There is no summary for this article')
+            updated = RSSNews.feed.get('updated', 'There is no summary for this article')
+
+            if summary != 'There is no summary for this article':
+
+                self.embed.add_field(name = f'{nr}. {article.title}', value = f'\n{summary}\n**{updated}**\n{article.link}\n')
+
+            if nr == 5:
+                break
+
+        #   Send the information, and reset embed
+        await ch.send(embed=self.embed)
+        self.embed.clear_fields()
+
+        return
+
+    @command(name='24europe')
+    async def France24Europe(self, ctx):
+
+            #   Retrieve the guild information
+        srv = ctx.guild
+        member = get(srv.roles, name='@Members')
+        ch = get(srv.channels, name=f'{self.channelName}')
+        rss = 'https://www.france24.com/en/europe/rss'
+
+        #   Creating the channel if it does not exists
+        if not ch:
+
+            #   Inserting the channel permissions
+            perms = {
+                        srv.default_role:PermissionOverwrite(read_messages=False),
+                        member:PermissionOverwrite(view=True, read_channel_history=True)
+}
+
+            await srv.create_text_channel(f'{self.channelName}', overwrites=perms)
+
+        #   Creating the RSS 
+        RSSNews = feedparser.parse(f'{rss}')
+        entries = RSSNews.entries
+
+        #   Prepareing the embed
+        self.embed.title = f'France 24 Europe'
+        self.embed.description = f''
+
+
+        #   Looping throught the entries
+        for nr, article in enumerate(entries):
+
+            summary = article.get('summary', 'There is no summary for this article')
+            updated = RSSNews.feed.get('updated', 'There is no summary for this article')
+
+            if summary != 'There is no summary for this article':
+
+                self.embed.add_field(name = f'{nr}. {article.title}', value = f'\n{summary}\n**{updated}**\n{article.link}\n')
+
+            if nr == 5:
+                break
+
+        #   Send the information, and reset embed
+        await ch.send(embed=self.embed)
+        self.embed.clear_fields()
+
+        return
+
+    @command(name='24africa')
+    async def France24Africa(self, ctx):
+
+            #   Retrieve the guild information
+        srv = ctx.guild
+        member = get(srv.roles, name='@Members')
+        ch = get(srv.channels, name=f'{self.channelName}')
+        rss = 'https://www.france24.com/en/africa/rss'
+
+        #   Creating the channel if it does not exists
+        if not ch:
+
+            #   Inserting the channel permissions
+            perms = {
+                        srv.default_role:PermissionOverwrite(read_messages=False),
+                        member:PermissionOverwrite(view=True, read_channel_history=True)
+}
+
+            await srv.create_text_channel(f'{self.channelName}', overwrites=perms)
+
+        #   Creating the RSS 
+        RSSNews = feedparser.parse(f'{rss}')
+        entries = RSSNews.entries
+
+        #   Prepare the embed
+        self.embed.title = f'France 24 Africa'
+        self.embed.description = f''
+        
+
+
+        #   Looping throught the entries
+        for nr, article in enumerate(entries):
+
+            summary = article.get('summary', 'There is no summary for this article')
+            updated = RSSNews.feed.get('updated', 'There is no summary for this article')
+
+            if summary != 'There is no summary for this article':
+
+                self.embed.add_field(name = f'{nr}. {article.title}', value = f'\n{summary}\n**{updated}**\n{article.link}\n')
+
+            if nr == 5:
+                break
+
+        #   Send the information, and reset embed
+        await ch.send(embed=self.embed)
+        self.embed.clear_fields()
+
+        return
+
+    @command(name='24me')
+    async def France24MiddleEast(self, ctx):
+
+            #   Retrieve the guild information
+        srv = ctx.guild
+        member = get(srv.roles, name='@Members')
+        ch = get(srv.channels, name=f'{self.channelName}')
+        rss = 'https://www.france24.com/en/middle-east/rss'
+
+        #   Creating the channel if it does not exists
+        if not ch:
+
+            #   Inserting the channel permissions
+            perms = {
+                        srv.default_role:PermissionOverwrite(read_messages=False),
+                        member:PermissionOverwrite(view=True, read_channel_history=True)
+}
+
+            await srv.create_text_channel(f'{self.channelName}', overwrites=perms)
+
+        #   Creating the RSS 
+        RSSNews = feedparser.parse(f'{rss}')
+        entries = RSSNews.entries
+
+        #   Prepare the embed
+        self.embed.title = f'France 24 Middle East'
+        self.embed.description = f''
+
+        #   Looping throught the entries
+        for nr, article in enumerate(entries):
+
+            summary = article.get('summary', 'There is no summary for this article')
+            updated = RSSNews.feed.get('updated', 'There is no summary for this article')
+
+            if summary != 'There is no summary for this article':
+
+                self.embed.add_field(name = f'{nr}. {article.title}', value = f'\n{summary}\n**{updated}**\n{article.link}\n')
+
+            if nr == 5:
+                break
+
+        #   Send the information, and reset embed
+        await ch.send(embed=self.embed)
+        self.embed.clear_fields()
+
+        return
+
+    @command(name='24america')
+    async def France24America(self, ctx):
+
+            #   Retrieve the guild information
+        srv = ctx.guild
+        member = get(srv.roles, name='@Members')
+        ch = get(srv.channels, name=f'{self.channelName}')
+        rss = 'https://www.france24.com/en/americas/rss'
+
+        #   Creating the channel if it does not exists
+        if not ch:
+
+            #   Inserting the channel permissions
+            perms = {
+                        srv.default_role:PermissionOverwrite(read_messages=False),
+                        member:PermissionOverwrite(view=True, read_channel_history=True)
+}
+
+            await srv.create_text_channel(f'{self.channelName}', overwrites=perms)
+
+        #   Creating the RSS 
+        RSSNews = feedparser.parse(f'{rss}')
+        entries = RSSNews.entries
+
+        #   Prepare the embed
+        self.embed.title = f'France 24 America'
+        self.embed.description = f''
+
+        #   Looping throught the entries
+        for nr, article in enumerate(entries):
+
+            summary = article.get('summary', 'There is no summary for this article')
+            updated = RSSNews.feed.get('updated', 'There is no summary for this article')
+
+            if summary != 'There is no summary for this article':
+
+                self.embed.add_field(name = f'{nr}. {article.title}', value = f'\n{summary}\n**{updated}**\n{article.link}\n')
+
+            if nr == 5:
+                break
+
+        #   Send the information, and reset embed
+        await ch.send(embed=self.embed)
+        self.embed.clear_fields()
+
+        return
+
+    @command(name='24asia')
+    async def France24Asia(self, ctx):
+
+            #   Retrieve the guild information
+        srv = ctx.guild
+        member = get(srv.roles, name='@Members')
+        ch = get(srv.channels, name=f'{self.channelName}')
+        rss = 'https://www.france24.com/en/asia-pacific/rss'
+
+        #   Creating the channel if it does not exists
+        if not ch:
+
+            #   Inserting the channel permissions
+            perms = {
+                        srv.default_role:PermissionOverwrite(read_messages=False),
+                        member:PermissionOverwrite(view=True, read_channel_history=True)
+}
+
+            await srv.create_text_channel(f'{self.channelName}', overwrites=perms)
+
+        #   Creating the RSS 
+        RSSNews = feedparser.parse(f'{rss}')
+        entries = RSSNews.entries
+
+        #   Prepare embed
+        self.embed.title = f'France 24 Asia'
+        self.embed.description = f''
+
+        #   Looping throught the entries
+        for nr, article in enumerate(entries):
+
+            summary = article.get('summary', 'There is no summary for this article')
+            updated = RSSNews.feed.get('updated', 'There is no summary for this article')
+
+            if summary != 'There is no summary for this article':
+
+                self.embed.add_field(name = f'{nr}. {article.title}', value = f'\n{summary}\n**{updated}**\n{article.link}\n')
+
+            if nr == 5:
+                break
+
+        #   Send the information, and reset embed
+        await ch.send(embed=self.embed)
+        self.embed.clear_fields()
+
+        return
+